Move coastline geometry out of the tide join
Two joins need great-circle distance now - a beach to a tide station, and a beach
to a wave buoy - and geometry filed under either would be found by whoever was
not looking for it. It moves verbatim into scripts/geo.mjs, with its tests, and
the tide join imports it.

No behaviour changes: the functions are unaltered, and the seeded inventory is
byte-identical.

Verified with npm run gate: format, lint, typecheck, test, build, stylesheet all
pass.
