+++ src/lib/__fixtures__/ndbc-46254-realtime2.txt
@@ -1,0 +1,8 @@
+#YY  MM DD hh mm WDIR WSPD GST  WVHT   DPD   APD MWD   PRES  ATMP  WTMP  DEWP  VIS PTDY  TIDE
+#yr  mo dy hr mn degT m/s  m/s     m   sec   sec degT   hPa  degC  degC  degC  nmi  hPa    ft
+2026 08 18 03 26  MM   MM   MM   0.8     5   4.6 278     MM    MM  21.1    MM   MM   MM    MM
+2026 08 18 02 56  MM   MM   MM   0.8     5   4.8 275     MM    MM  21.1    MM   MM   MM    MM
+2026 08 18 02 26  MM   MM   MM   0.8     4   4.6 273     MM    MM  21.2    MM   MM   MM    MM
+2026 08 18 01 56  MM   MM   MM   0.7     5   4.7 282     MM    MM  21.4    MM   MM   MM    MM
+2026 08 18 01 26  MM   MM   MM   0.7     6   4.6 294     MM    MM  21.5    MM   MM   MM    MM
+2026 08 18 00 56  MM   MM   MM   0.7     6   4.7 288     MM    MM  21.6    MM   MM   MM    MM
